Move trace test to a separate file.
